+++ lib/active_record/connection_adapters/sqlserver/type.rb
@@ -1,5 +1,4 @@
 require 'active_record/type'
-require 'active_record/connection_adapters/sqlserver/type/quoter.rb'
 # Exact Numerics
 require 'active_record/connection_adapters/sqlserver/type/integer.rb'
 require 'active_record/connection_adapters/sqlserver/type/big_integer.rb'
@@ -13,8 +12,11 @@ require 'active_record/connection_adapters/sqlserver/type/small_money.rb'
 require 'active_record/connection_adapters/sqlserver/type/float.rb'
 require 'active_record/connection_adapters/sqlserver/type/real.rb'
 # Date and Time
+require 'active_record/connection_adapters/sqlserver/type/time_value_fractional.rb'
 require 'active_record/connection_adapters/sqlserver/type/date.rb'
 require 'active_record/connection_adapters/sqlserver/type/datetime.rb'
+require 'active_record/connection_adapters/sqlserver/type/datetime2.rb'
+require 'active_record/connection_adapters/sqlserver/type/datetimeoffset.rb'
 require 'active_record/connection_adapters/sqlserver/type/smalldatetime.rb'
 require 'active_record/connection_adapters/sqlserver/type/time.rb'
 # Character Strings